backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<0.31.0
SQLAlchemy>=2.0.30,<2.1.0
psycopg2-binary>=2.9.9,<3.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
dnspython
python-whois
requests
beautifulsoup4
python-nmap==0.7.1
dnspython==2.7.0